fix: build errors due to undefined properties
