+++ packages/core/src/lib/storage/index.ts
@@ -1,2 +1,3 @@
+export * from './storage';
 export * from './storage.service';
 export * from './storage.interface';
